requirements.txt
requests
python-dotenv
markdown
